fix: move reward to types
